services/listening_engine/requirements.txt
# Termómetro Cultural - Social Sentiment Monitoring System
# Python 3.11+

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Scraping
playwright>=1.41.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
httpx>=0.26.0

# NLP / LLM
openai>=1.66.0
# grok / x.ai client if needed

# Queue / Cache / Scheduler
redis>=5.0.0
celery[redis]>=5.3.0        # Celery with Redis transport
kombu>=5.3.0                 # Celery message library (routing / queues)
# flower>=2.0.0              # Optional: Celery monitoring UI (celery -A ... flower)

# Config & validation
pyyaml>=6.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utils
python-dateutil>=2.8.2
structlog>=24.1.0
tenacity>=8.2.0

# Apify (optional)
# apify-client>=1.6.0

# Dev / Test
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
